answer-service/requirements.txt
nats-py==2.12.0
dspy==3.0.4
python-dotenv==1.0.1
